refactor: move edge config to config crate
